feat: rename AccountNavigation component to TopNavigation
